fix(dsp): improve EQ bypass transitions to prevent audible pops fix(ui): refactor TrackTable to simplify scroll view creation fix(ui): streamline TrackTableCoordinator's data handling fix(ui): enhance PlaylistSidebarViewModel to handle missing nodes gracefully test(ui): add snapshot tests for IdentifyTrackSheet states test(ui): improve SmartPlaylistDetailViewModelTests with proper async handling
