+++ Modules/UI/Tests/UITests/ViewModelTests/SmartPlaylistDetailViewModelTests.swift
@@ -68,9 +68,12 @@ struct SmartPlaylistDetailViewModelTests {
         await seed.vm.load(playlistID: seed.playlistID)
 
         // Allow the observation stream to emit the initial value.
+        // Task.yield() alone is insufficient — GRDB observation involves async
+        // database I/O before the first emission, so a short sleep per iteration
+        // lets those operations complete while the MainActor is suspended.
         for _ in 0 ..< 50 {
             if !seed.vm.tracks.isEmpty { break }
-            await Task.yield()
+            try await Task.sleep(nanoseconds: 10_000_000) // 10 ms per iteration, 500 ms max
         }
 
         #expect(seed.vm.title == "Loved")
